Move multi-queue concurrency test demo out of main class and into test/
